fix portable cell not working
