+++ src/main/java/appeng/container/implementations/ContainerWirelessTerm.java
@@ -23,7 +23,7 @@ import appeng.helpers.WirelessTerminalGuiObject;
 import net.minecraft.entity.player.InventoryPlayer;
 
 
-public class ContainerWirelessTerm extends ContainerMEPortableCell {
+public class ContainerWirelessTerm extends ContainerMEPortableTerminal {
 
     public ContainerWirelessTerm(final InventoryPlayer ip, final WirelessTerminalGuiObject gui) {
         super(ip, gui, true);
